Refactor repo to follow standards and update to FG4
Also, Kotlin build script rather than Groovy so that IntelliJ has a clue
about what's written in it :stuck_out_tongue:
