+++ .gitignore
@@ -1,0 +1,29 @@
+# eclipse
+bin
+*.launch
+.settings
+.metadata
+.classpath
+.project
+
+# idea
+out
+*.ipr
+*.iws
+*.iml
+.idea
+
+# gradle
+build
+.gradle
+
+# other
+eclipse
+run_client
+run_server
+
+# Files from Forge MDK
+forge*changelog.txt
+
+# Java stuff
+*.jar
