--- DimensionHopperFinalBoss/src/main/java/com/TheComputerizer/DimensionalHopperFinalBoss/Reference.java
@@ -1,13 +1,0 @@
-package com.TheComputerizer.DimensionalHopperFinalBoss;
-
-public class Reference {
-    public static final String MOD_ID = "dimensionalhopperfinalboss";
-    public static final String VERSION = "1.0";
-    public static final String MC_VERSION = "[1.12.2]";
-    public static final String NAME = "Dimension Hopper Final Boss";
-
-    public static final String CLIENT_PROXY_CLASS = "com.TheComputerizer.DimensionalHopperFinalBoss.proxy.ClientProxy";
-    public static final String COMMON_PROXY_CLASS = "com.TheComputerizer.DimensionalHopperFinalBoss.proxy.CommonProxy";
-
-    public static final int ENTITY_FINAL_BOSS = 120;
-}
