--- DimensionHopperFinalBoss/src/main/java/com/TheComputerizer/DimensionalHopperFinalBoss/proxy/CommonProxy.java
@@ -1,11 +1,0 @@
-package com.TheComputerizer.DimensionalHopperFinalBoss.proxy;
-
-import net.minecraft.item.Item;
-
-public class CommonProxy
-{
-    public void registerItemRenderer(Item item, int meta, String id)
-    {
-
-    }
-}
