+++ src/main/java/com/TheComputerizer/DimensionalHopperFinalBoss/proxy/ClientProxy.java
@@ -1,0 +1,14 @@
+package com.TheComputerizer.DimensionalHopperFinalBoss.proxy;
+
+import net.minecraft.client.renderer.block.model.ModelResourceLocation;
+import net.minecraft.item.Item;
+import net.minecraftforge.client.model.ModelLoader;
+
+public class ClientProxy extends CommonProxy
+{
+    @Override
+    public void registerItemRenderer(Item item, int meta, String id)
+    {
+        ModelLoader.setCustomModelResourceLocation(item, meta, new ModelResourceLocation(item.getRegistryName(), id));
+    }
+}
