+++ src/main/java/com/TheComputerizer/DimensionalHopperFinalBoss/util/handlers/RegistryHandler.java
@@ -1,0 +1,57 @@
+package com.TheComputerizer.DimensionalHopperFinalBoss.util.handlers;
+
+
+import com.TheComputerizer.DimensionalHopperFinalBoss.init.EntityInit;
+import com.TheComputerizer.DimensionalHopperFinalBoss.init.ItemInit;
+import com.TheComputerizer.DimensionalHopperFinalBoss.util.interfaces.IHasModel;
+import net.minecraft.enchantment.Enchantment;
+import net.minecraft.entity.Entity;
+import net.minecraft.item.Item;
+import net.minecraftforge.client.event.ModelRegistryEvent;
+import net.minecraftforge.event.RegistryEvent;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
+import net.minecraftforge.fml.common.event.FMLServerStartingEvent;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
+
+@EventBusSubscriber
+public class RegistryHandler
+{
+    @SubscribeEvent
+    public static void onItemRegister(RegistryEvent.Register<Item> event)
+    {
+        event.getRegistry().registerAll(ItemInit.ITEMS.toArray(new Item[0]));
+    }
+
+    @SubscribeEvent
+    public static void onModelRegister(ModelRegistryEvent event)
+    {
+        for(Item item : ItemInit.ITEMS)
+        {
+            if(item instanceof IHasModel)
+            {
+                ((IHasModel)item).registerModels();
+            }
+        }
+    }
+
+    public static void preInitRegistries()
+    {
+        EntityInit.registerEntities();
+    }
+
+    public static void initRegistries()
+    {
+
+    }
+
+    public static void postInitRegistries()
+    {
+
+    }
+
+    public static void serverRegistries(FMLServerStartingEvent event)
+    {
+
+    }
+}
