+++ src/main/java/com/TheComputerizer/DimensionalHopperFinalBoss/util/interfaces/IHasModel.java
@@ -1,0 +1,5 @@
+package com.TheComputerizer.DimensionalHopperFinalBoss.util.interfaces;
+
+public interface IHasModel {
+    public void registerModels();
+}
