+++ src/main/resources/assets/dimensionalhopperfinalboss/textures/entity/finalboss.png
