+++ src/main/resources/mcmod.info
