+++ src/main/resources/pack.mcmeta
